Move out some functions from Language.Finkel.Fnk
Move command line option related functions in Language.Finkel.Fnk to
Language.Finkel.Options. Add Language.Finkel.Options to the list of exposed
module in cabal configuration.

Move printFinkelException in Language.Finkel.Fnk to Language.Finkel.Exception.

Modify the imports in the modules using the moved functions.
