+++ finkel-kernel/src/Language/Finkel/Exception.hs
@@ -1,4 +1,5 @@
-{-# LANGUAGE CPP #-}
+{-# LANGUAGE CPP              #-}
+{-# LANGUAGE TypeApplications #-}
 
 -- | Exception related types and functions in @finkel-kernel@.
 module Language.Finkel.Exception
@@ -6,26 +7,50 @@ module Language.Finkel.Exception
   , finkelExceptionLoc
   , readOrFinkelException
   , handleFinkelException
+  , printFinkelException
   ) where
 
 #include "ghc_modules.h"
 
 -- base
-import Control.Exception    (Exception (..), throw)
+import Control.Exception            (Exception (..), throw)
+import Control.Monad.IO.Class       (MonadIO (..))
+import System.IO                    (hPutStrLn, stderr)
 
 -- ghc
-import GHC_Types_SrcLoc     (GenLocated (..), SrcSpan)
-import GHC_Utils_Exception  (ExceptionMonad)
+import GHC_Data_Bag                 (unitBag)
+import GHC_Driver_Session           (HasDynFlags (..))
+import GHC_Types_SrcLoc             (GenLocated (..), SrcSpan)
+import GHC_Utils_Exception          (ExceptionMonad)
+import GHC_Utils_Outputable         (neverQualify, text)
+
+#if MIN_VERSION_ghc(9,6,0)
+import GHC.Driver.Errors.Types      (GhcMessage)
+import GHC.Types.Error              (defaultDiagnosticOpts)
+#endif
+
+#if MIN_VERSION_ghc(9,4,0)
+import GHC.Driver.Config.Diagnostic (initDiagOpts)
+import GHC.Driver.Errors            (printMessages)
+import GHC.Types.Error              (mkMessages)
+#else
+import GHC_Driver_Errors            (printBagOfErrors)
+#endif
+
+#if MIN_VERSION_ghc(9,2,0)
+import GHC.Utils.Logger             (HasLogger (..))
+#endif
 
 #if MIN_VERSION_ghc(9,0,0)
 -- exceptions
-import Control.Monad.Catch  (handle)
+import Control.Monad.Catch          (handle)
 #else
 -- ghc
-import GHC_Utils_Exception  (ghandle)
+import GHC_Utils_Exception          (ghandle)
 #endif
 
 -- Internal
+import Language.Finkel.Error
 import Language.Finkel.Form
 
 
@@ -76,6 +101,41 @@ readOrFinkelException what name str =
                                  " for " ++ name ++
                                  ", but got " ++ show str))
 {-# INLINABLE readOrFinkelException #-}
+
+-- | Print 'FinkelException' with source code information when available.
+#if MIN_VERSION_ghc(9,2,0)
+printFinkelException
+  :: (HasLogger m, HasDynFlags m, MonadIO m) => FinkelException -> m ()
+#else
+printFinkelException :: (HasDynFlags m, MonadIO m) => FinkelException -> m ()
+#endif
+printFinkelException e = case finkelExceptionLoc e of
+  Just l  -> prLocErr l
+  Nothing -> pr msg
+  where
+    msg = displayException e
+    pr = liftIO . hPutStrLn stderr
+    prLocErr l = do
+      dflags <- getDynFlags
+      let em = mkWrappedMsg dflags l neverQualify (text msg)
+#if MIN_VERSION_ghc(9,6,0)
+      logger <- getLogger
+      let ghc_msg = mkMessages (unitBag em)
+          diagnostic_opts = defaultDiagnosticOpts @GhcMessage
+          diag_opts = initDiagOpts dflags
+      liftIO (printMessages logger diagnostic_opts diag_opts ghc_msg)
+#elif MIN_VERSION_ghc(9,4,0)
+      logger <- getLogger
+      let ghc_msg = mkMessages (unitBag em)
+      liftIO (printMessages logger (initDiagOpts dflags) ghc_msg)
+#elif MIN_VERSION_ghc(9,2,0)
+      logger <- getLogger
+      liftIO (printBagOfErrors logger dflags (unitBag em))
+#else
+      liftIO (printBagOfErrors dflags (unitBag em))
+#endif
+{-# INLINABLE printFinkelException #-}
+
 
 -- ------------------------------------------------------------------------
 --
